Simulation support for shooter
This commit adds simulation support for the shooter. The shooter uses
PID on board the SparkMAX. We copy the SparkMAX onboard implementation
in Java, using documentation from Rev. We use inheritance to replace
instances of SparkPIDController with SparkPIDControllerSim, and then
pass velocity from the physics simulation to the PID sim each step. In
turn, we use the PID sim to update the voltage for the physics sim.