mostly functioning aim and rotate code, mostly by logan
There's still an issue occasionally when there is no target or it's blinking a lot? I think it might be an issue of the target being lost between function calls to check for a target and use its data, but that seems to be unlikely. Nonetheless, this is an issue I'm having problems reproducing.